feat(chart): bump wordpress chart version to 0.8.12, change PV reclaim policy to Recycle
This commit updates the Wordpress Helm chart by incrementing the chart version following a change in the volume configuration. The `persistentVolumeReclaimPolicy` is modified from `Delete` to `Recycle`, allowing for the reuse of the volume across different claims. This can be useful to retain data without the need to back it up externally before redeployment. Be aware that recycling is deprecated in Kubernetes v1.16+, and this option might not be available in future versions.